refactor: 테스트 LetterLogService 생성
