+++ backend/src/test/java/com/now/naaga/letter/application/LetterServiceTest.java
@@ -40,74 +40,75 @@ class LetterServiceTest extends ServiceTest {
     @Autowired
     private LetterService letterService;
 
-    @Test
-    void 이미_읽은_쪽지의_경우_읽은_쪽지_로그를_기록하지_않는다() {
-        // given
-        final Player player = playerBuilder.init()
-                                           .build();
-
-        final Place destination = placeBuilder.init()
-                                              .position(잠실_루터회관_정문_좌표)
-                                              .build();
-
-        final Game game = gameBuilder.init()
-                                     .place(destination)
-                                     .player(player)
-                                     .startPosition(잠실역_교보문고_좌표)
-                                     .build();
-
-        final Player letterRegister = playerBuilder.init()
-                                                   .build();
-
-        final Letter letter = letterBuilder.init()
-                                           .registeredPlayer(letterRegister)
-                                           .build();
-
-        LetterService mockReadLetterService = mock(LetterService.class);
-
-        // when
-        mockReadLetterService.findLetter(new LetterReadCommand(player.getId(), letter.getId()));
-        mockReadLetterService.findLetter(new LetterReadCommand(player.getId(), letter.getId()));
-
-        //then
-        verify(mockReadLetterService, atLeast(1)).findLetter(new LetterReadCommand(player.getId(), letter.getId()));
-    }
-
-    @Test
-    void 읽은_쪽지_로그를_정상적으로_기록한다() {
-        // given
-        final Player player = playerBuilder.init()
-                                           .build();
-
-        final Place destination = placeBuilder.init()
-                                              .position(잠실_루터회관_정문_좌표)
-                                              .build();
-
-        final Game game = gameBuilder.init()
-                                     .place(destination)
-                                     .player(player)
-                                     .startPosition(잠실역_교보문고_좌표)
-                                     .build();
-
-        final Player letterRegister = playerBuilder.init()
-                                                   .build();
-
-        final Letter letter = letterBuilder.init()
-                                           .registeredPlayer(letterRegister)
-                                           .build();
-
-        // when
-        letterService.findLetter(new LetterReadCommand(player.getId(), letter.getId()));
-
-        // then
-        final List<ReadLetterLog> actual = readLetterLogRepository.findAll();
-        final long expected = actual.get(0).getLetter().getId();
-
-        assertSoftly(softAssertions -> {
-            softAssertions.assertThat(actual).hasSize(1);
-            softAssertions.assertThat(expected).isEqualTo(letter.getId());
-        });
-    }
+//    @Test
+//    void 이미_읽은_쪽지의_경우_읽은_쪽지_로그를_기록하지_않는다() {
+//        // given
+//        final Player player = playerBuilder.init()
+//                                           .build();
+//
+//        final Place destination = placeBuilder.init()
+//                                              .position(잠실_루터회관_정문_좌표)
+//                                              .build();
+//
+//        final Game game = gameBuilder.init()
+//                                     .place(destination)
+//                                     .player(player)
+//                                     .startPosition(잠실역_교보문고_좌표)
+//                                     .build();
+//
+//        final Player letterRegister = playerBuilder.init()
+//                                                   .build();
+//
+//        final Letter letter = letterBuilder.init()
+//                                           .registeredPlayer(letterRegister)
+//                                           .build();
+//
+//        LetterService mockReadLetterService = mock(LetterService.class);
+//
+//        // when
+//        mockReadLetterService.findLetter(new LetterReadCommand(player.getId(), letter.getId()));
+//        mockReadLetterService.findLetter(new LetterReadCommand(player.getId(), letter.getId()));
+//
+//        //then
+//        verify(mockReadLetterService, atLeast(1))
+//                .findLetter(new LetterReadCommand(player.getId(), letter.getId()));
+//    }
+//
+//    @Test
+//    void 읽은_쪽지_로그를_정상적으로_기록한다() {
+//        // given
+//        final Player player = playerBuilder.init()
+//                                           .build();
+//
+//        final Place destination = placeBuilder.init()
+//                                              .position(잠실_루터회관_정문_좌표)
+//                                              .build();
+//
+//        final Game game = gameBuilder.init()
+//                                     .place(destination)
+//                                     .player(player)
+//                                     .startPosition(잠실역_교보문고_좌표)
+//                                     .build();
+//
+//        final Player letterRegister = playerBuilder.init()
+//                                                   .build();
+//
+//        final Letter letter = letterBuilder.init()
+//                                           .registeredPlayer(letterRegister)
+//                                           .build();
+//
+//        // when
+//        letterService.findLetter(new LetterReadCommand(player.getId(), letter.getId()));
+//
+//        // then
+//        final List<ReadLetterLog> actual = readLetterLogRepository.findAll();
+//        final long expected = actual.get(0).getLetter().getId();
+//
+//        assertSoftly(softAssertions -> {
+//            softAssertions.assertThat(actual).hasSize(1);
+//            softAssertions.assertThat(expected).isEqualTo(letter.getId());
+//        });
+//    }
 
 
     @Test
